watched the positioning video
Why is the footer not getting in line with the rest of the container? It says it is 1200px however it is falling slightly outside the line.

I am having trouble with the Features section text. The space between the paragraph lines looks too small.

Generally on figma, I have no tool like you do to measure pixels.

Is it ok the way I put the classes for the portfolio images? To get the space between the cards excluding every last one in the row, I did not include a class on that card. There might be a better way to do this with .portfolio-images:not:nth-child(3n+3)...